content_posts/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
python-decouple
dj-database-url
whitenoise
requests
gunicorn
psycopg2-binary